refactor(frontend): move component to dedicate file
