+++ apps/frontend/app/components/common.tsx
@@ -70,6 +70,7 @@ import {
 	IconMoodHappy,
 	IconMoodSad,
 	IconRefresh,
+	IconRotateClockwise,
 	IconScaleOutline,
 	IconSearch,
 	IconServer,
@@ -1356,5 +1357,36 @@ export const DisplayListDetailsAndRefresh = (props: {
 				</Button>
 			</Form>
 		</Group>
+	);
+};
+
+export const ExpireCacheKeyButton = (props: {
+	cacheId: string;
+	confirmationText?: string;
+}) => {
+	const submit = useConfirmSubmit();
+
+	return (
+		<Form
+			replace
+			method="POST"
+			action={withQuery($path("/actions"), { intent: "expireCacheKey" })}
+		>
+			<input type="hidden" name="cacheId" value={props.cacheId} />
+			<ActionIcon
+				type="submit"
+				variant="subtle"
+				onClick={(e) => {
+					if (!props.confirmationText) return;
+					const form = e.currentTarget.form;
+					if (form) {
+						e.preventDefault();
+						openConfirmationModal(props.confirmationText, () => submit(form));
+					}
+				}}
+			>
+				<IconRotateClockwise />
+			</ActionIcon>
+		</Form>
 	);
 };
